// apps/sync-service/src/openapi/order-fulfilled.ts
export type PutV2OrderIdFulfilledBodyOrderItemsItemOrderStockFulfillItem = {
  order_stock_ids: number[];
  /** @nullable */
  batch_id: number | null;
  /** @nullable */
  status: number | null;
  fulfill_reason: null;
  other_reason: null;
  /** @nullable */
  qrcode: string | null;
  received_qty: number;
};

export type PutV2OrderIdFulfilledBodyOrderItemsItem = {
  id: number;
  material_id: number;
  order_stock_fulfill: PutV2OrderIdFulfilledBodyOrderItemsItemOrderStockFulfillItem[];
};

export type PutV2OrderIdFulfilledBody = {
  comment: string;
  fulfilled_at: string;
  order_items: PutV2OrderIdFulfilledBodyOrderItemsItem[];
};